backend/phone_recs_bot/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
motor==3.6.0
pymongo==4.9.2